Rename model package to validated_package
